requirements.txt
python-telegram-bot==20.0a4  # Substitua pela versão que você está usando
